Fix some issues caused by config being too large

+++ src/avr/lib/eeprom/eeprom.c
@@ -51,10 +51,10 @@ void loadConfig(void) {
     eeprom_update_block(&config, &config_pointer, sizeof(Configuration_t));
   }
 }
-void writeConfigByte(uint8_t offset, uint8_t byte) {
+void writeConfigByte(uint16_t offset, uint8_t byte) {
   eeprom_update_byte(((uint8_t *)&config_pointer) + offset, byte);
 }
-void writeConfigBlock(uint8_t offset, const uint8_t* data, uint16_t len) {
+void writeConfigBlock(uint16_t offset, const uint8_t* data, uint16_t len) {
   eeprom_update_block(data, ((uint8_t *)&config_pointer) + offset, len);
 }
